dev-requirements.txt
-r requirements.txt
ddt==1.6.0
django-debug-toolbar==3.8.1
django-silk==5.0.2
factory-boy==3.2.1
Faker==15.3.2
pytest==7.2.0
pytest-django==4.5.2
pre-commit==2.21.0
pytest-xdist==3.2.0
pytest-cov==4.0.0
flake8==6.0.0
black==22.12.0
isort==5.11.5
freezegun==1.2.2
requests==2.30.0